Removing warnings from NodeBase file

+++ include/geos/index/quadtree/NodeBase.h
@@ -83,7 +83,7 @@ public:
 
 	size_t size() const;
 
-	unsigned int getNodeCount() const;
+	size_t getNodeCount() const;
 
 	virtual std::string toString() const;
